(refactor) Moved app into server in prep for AWS deployment.
